requirements.txt
APScheduler==3.6.3
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.7
cryptography==3.2
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iocextract==1.13.1
iocparser==1.0.14
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
moviepy==1.0.3
numpy==1.19.4
pdfminer==20191125
Pillow==8.0.1
pip==20.2.3
proglog==0.1.9
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
pylint==2.6.0
PyPDF2==1.26.0
python-telegram-bot==13.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
setuptools==49.2.1
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tornado==6.0.4
tqdm==4.52.0
tzlocal==2.1
urllib3==1.25.11
waitress==1.4.4
Werkzeug==1.0.1
wheel==0.35.1
wrapt==1.12.1